+++ .vscode/settings.json
@@ -2,6 +2,15 @@
     "files.associations": {
         "*.txx": "cpp",
         "chrono": "cpp",
-        "*.in": "cpp"
+        "*.in": "cpp",
+        "fstream": "cpp",
+        "istream": "cpp",
+        "memory": "cpp",
+        "new": "cpp",
+        "ostream": "cpp",
+        "sstream": "cpp",
+        "type_traits": "cpp",
+        "core": "cpp",
+        "svd": "cpp"
     }
 }
