+++ Applications/shapeworks/Command.h
@@ -79,6 +79,14 @@ public:
 private:
 };
 
+class LdsAnalyzeCommandGroup : public Command
+{
+public:
+  const std::string type() override { return "LdsAnalyze"; }
+
+private:
+};
+
 class ParticleSystemCommand : public Command
 {
 public:
