+++ Applications/shapeworks/Commands.cpp
@@ -176,4 +176,72 @@ bool GroomCommand::execute(const optparse::Values& options, SharedCommandData& s
   }
 }
 
+
+///////////////////////////////////////////////////////////////////////////////
+// LdsAnalyze
+///////////////////////////////////////////////////////////////////////////////
+void LdsAnalyzeCommand::buildParser()
+{
+  const std::string prog = "ldsanalyze";
+  const std::string desc = "Analyze a particle system through LDS Model";
+  parser.prog(prog).description(desc);
+
+  parser.add_option("--name").action("store").type("string").set_default("").help("Path to parameter file.");
+
+  Command::buildParser();
+}
+
+bool LdsAnalyzeCommand::execute(const optparse::Values &options, SharedCommandData &sharedData)
+{
+  const std::string& projectFile(static_cast<std::string>(options.get("name")));
+
+  if (projectFile.length() == 0)
+  {
+    std::cerr << "Must specify project name\n";
+    return false;
+  }
+
+  bool isProject = StringUtils::hasSuffix(projectFile, "xlsx");
+
+  Optimize app;
+  if (isProject) {
+    try {
+      // load spreadsheet project
+      ProjectHandle project = std::make_shared<Project>();
+      project->load(projectFile);
+
+      const auto oldBasePath = boost::filesystem::current_path();
+      auto base = StringUtils::getPath(projectFile);
+      if (base != projectFile) {
+        boost::filesystem::current_path(base.c_str());
+        project->set_filename(StringUtils::getFilename(projectFile));
+      }
+
+      // set up Optimize class based on project parameters
+      OptimizeParameters params(project);
+      params.set_up_optimize(&app);
+      app.SetProject(project);
+
+      bool success = app.Run();
+
+      boost::filesystem::current_path(oldBasePath);
+      if (success) {
+        project->save(projectFile);
+      }
+
+      return success;
+    }
+    catch (std::exception& e) {
+      std::cerr << "Error: " << e.what() << "\n";
+      return false;
+    }
+  }
+  else {
+    OptimizeParameterFile param;
+    param.load_parameter_file(projectFile.c_str(), &app);
+    return app.Run();
+  }
+}
+
+///////////////////////////////////////////////////////////////////////////////
 } // shapeworks
