+++ Applications/shapeworks/shapeworks.cpp
@@ -104,6 +104,7 @@ int main(int argc, char *argv[])
   // Misc Commands
   shapeworks.addCommand(Seed::getCommand());
   shapeworks.addCommand(OptimizeCommand::getCommand());
+  shapeworks.addCommand(LdsAnalyzeCommand::getCommand());
   shapeworks.addCommand(GroomCommand::getCommand());
 
   try {
