+++ Libs/Particles/CMakeLists.txt
@@ -3,6 +3,7 @@
 set(Particles_sources
  ParticleSystem.cpp
  ParticleShapeStatistics.cpp
+ LdsAnalyze.cpp
  ShapeEvaluation.cpp
  ReconstructSurface.cpp
  ParticleNormalEvaluation.cpp
@@ -11,6 +12,7 @@ set(Particles_sources
 set(Particles_headers
  ParticleSystem.h
  ParticleShapeStatistics.h
+ LdsAnalyze.h
  itkParticlePositionReader.h
  itkParticlePositionWriter.h
  EvaluationUtil.h
