+++ Python/LDS_EM/build_models.py
@@ -1,0 +1,5 @@
+from Project_SSM import Project_SSM
+from Constants import DEFAULT_OPT_PARAMS
+
+proj_ob = Project_SSM(project_name = 'PRE_POST_MODEL')
+proj_ob.create_sw_project_file()
